FIX: Mouse having PrimaryAction on two buttons.
