+++ Packages/com.unity.inputsystem/InputSystem/Devices/Pointer.cs
@@ -23,10 +23,7 @@ namespace UnityEngine.Experimental.Input.LowLevel
     [StructLayout(LayoutKind.Sequential)]
     public struct PointerState : IInputStateTypeInfo
     {
-        public static FourCC kFormat
-        {
-            get { return new FourCC('P', 'T', 'R'); }
-        }
+        public static FourCC kFormat => new FourCC('P', 'T', 'R');
 
         [InputControl(layout = "Digital")]
         public uint pointerId;
