+++ Packages/com.unity.inputsystem/InputSystem/Devices/TouchManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20135adf5ac3b49a9b31f2aeb594bd64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
